requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
stripe==7.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
flask_cors==6.0.1
Flask-JWT-Extended===4.7.1
Flask-Migrate===4.1.0
redis===6.4.0
python-json-logger===4.0.0
